refactor(MilestoneService): Extract MilestoneCriteriaEvaluator
#1119
